+++ Appmeds/Appmeds/Medication.cs
@@ -7,8 +7,8 @@ namespace Appmeds
     public class Medication
     {
         public string MedicationName { get; set; }
-        public int Dosage { get; set; }
-        public int NumberOfPills { get; set; }
+        public double Dosage { get; set; }
+        public double NumberOfPills { get; set; }
         public TimeSpan Time { get; set; }
     }
 }
